you need to set the position of the sprites, since they are the thing being rendered

+++ tbSFML/tbSFML/FoodHandler.h
@@ -3,6 +3,7 @@
 #include "Food.h"
 #include <deque>
 #include <list>
+#include <utility> // for std::pair
 
 class FoodHandler
 {
@@ -15,7 +16,7 @@ private:
 
 public:
 	void SpawnFood(int spawnAmount);
-	sf::Vector2f GenerateRandomCoordinates(); // They must be inside the screen, and not on the snake
+	sf::Vector2i GenerateRandomCoordinates(); // They must be inside the screen, and not on the snake
 	void DrawFood(sf::RenderWindow& window);
 	void RemoveApple(Food food); // This will re-spawn the apple in a different position on the screen
 	void EnsureAmountOfFoodOnScreen(int spawnAmount); // This will ensure that the amount of food on the screen is stays the same, since it can sometimes less, due to how the snake moves, and the apples are spawned
